+++ changelog.md
@@ -2,6 +2,9 @@ Pending Next Release
 =========
 - Update to Microsoft.Identity.Abstractions 5.2.0
 
+### New features
+- Added support for Managed Identity Federated Identity Credential. See issue [2749](https://github.com/AzureAD/microsoft-identity-web/issues/2749) for details.
+
 2.17.2
 =========